requirements.txt
prokka		1.12
blast		2.5.0+
future		0.18.2
scikit-learn	0.23.2
biopython	1.78
numpy		1.19.2
pandas		1.1.3
torch		1.9.0+cpu
dominate		2.9.1